requirements.txt
aiohttp>=3.8.4
requests>=2.28.2
numpy>=1.24.3
python-dotenv>=1.0.0
pydantic>=2.4.2
pytest>=7.3.1
pytest-asyncio>=0.21.0
langchain>=0.1.0
langgraph>=0.0.16
langchain-google-genai>=0.0.3
websockets>=11.0.3
fastapi>=0.103.1
uvicorn>=0.23.2
faiss-cpu>=1.7.4
